+++ pippo-core/src/main/java/ro/pippo/core/route/WebjarsResourceHandler.java
@@ -15,6 +15,26 @@
  */
 package ro.pippo.core.route;
 
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import ro.pippo.core.PippoRuntimeException;
+import ro.pippo.core.util.ClasspathUtils;
+
+import java.io.File;
+import java.io.FileInputStream;
+import java.io.IOException;
+import java.io.InputStream;
+import java.net.URI;
+import java.net.URISyntaxException;
+import java.net.URL;
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Map;
+import java.util.Properties;
+import java.util.TreeMap;
+import java.util.jar.JarEntry;
+import java.util.jar.JarInputStream;
+
 /**
  * Serves Webjars classpath resources.
  *
@@ -22,12 +42,103 @@ package ro.pippo.core.route;
  */
 public class WebjarsResourceHandler extends ClasspathResourceHandler {
 
+    private final Logger log = LoggerFactory.getLogger(WebjarsResourceHandler.class);
+
+    private final Map<String, String> pathAliases;
+
     public WebjarsResourceHandler() {
         this("/webjars");
     }
 
     public WebjarsResourceHandler(String urlPath) {
         super(urlPath, "META-INF/resources/webjars");
+        this.pathAliases = indexWebjars();
+    }
+
+    @Override
+    public URL getResourceUrl(String resourcePath) {
+        String resourceName;
+        String artifactPath = resourcePath.substring(0, resourcePath.indexOf('/', resourcePath.indexOf('/') + 1) + 1);
+        if (pathAliases.containsKey(artifactPath)) {
+            String artifactVersion = pathAliases.get(artifactPath);
+            String aliasedPath = resourcePath.replace(artifactPath, artifactVersion);
+            log.trace("Replaced Webjar path {} with {}", resourcePath, aliasedPath);
+            resourceName = getResourceBasePath() + "/" + aliasedPath;
+        } else {
+            resourceName = getResourceBasePath() + "/" + resourcePath;
+        }
+
+        URL url = this.getClass().getClassLoader().getResource(resourceName);
+        if (url == null) {
+            log.warn("Resource '{}' not found", resourceName);
+        }
+
+        return url;
+    }
+
+    /**
+     * Scans the classpath for Webjars metadata and creates an alias registry of "current" path to "fixed" path.
+     * <p>
+     * This allows specifications of a path like "jquery/current/jquery.min.js" which will be substituted
+     * at runtime as "jquery/1.11.1/jquery.min.js".</p>
+     *
+     * @return a map of path pathAliases
+     */
+    private Map<String, String> indexWebjars() {
+        Map<String, String> pathAliases = new TreeMap<>();
+        for (String pomProperties : locatePomProperties()) {
+            log.debug("Parsing Webjars metadata {}", pomProperties);
+            URL url = ClasspathUtils.locateOnClasspath(pomProperties);
+            Properties pom = new Properties();
+            try (InputStream is = url.openStream()) {
+                pom.load(is);
+            } catch (IOException e) {
+                log.error("Failed to read {}", url);
+                continue;
+            }
+
+            String artifactId = pom.getProperty("artifactId");
+            String version = pom.getProperty("version");
+            String fixedPath = artifactId + '/' + version + '/';
+            String currentPath = artifactId + "/current/";
+            pathAliases.put(currentPath, fixedPath);
+        }
+
+        for (Map.Entry<String, String> alias : pathAliases.entrySet()) {
+            log.debug("Registered Webjars path alias '{}' => '{}'", alias.getKey(), alias.getValue());
+        }
+        return pathAliases;
+    }
+
+    /**
+     * Locate all Webjars Maven pom.properties files on the classpath.
+     *
+     * @return a list of Maven pom.properties files on the classpath
+     */
+    private List<String> locatePomProperties() {
+        List<String> propertiesFiles = new ArrayList<>();
+        List<URL> packageUrls = ClasspathUtils.getResources(getResourceBasePath());
+        for (URL packageUrl : packageUrls) {
+            if (packageUrl.getProtocol().equals("jar")) {
+                // We only care about Webjars jar files
+                log.debug("Scanning {}", packageUrl);
+                try {
+                    String jar = packageUrl.toString().substring("jar:".length()).split("!")[0];
+                    File file = new File(new URI(jar));
+                    try (JarInputStream is = new JarInputStream(new FileInputStream(file))) {
+                        JarEntry entry = null;
+                        while ((entry = is.getNextJarEntry()) != null) {
+                            if (!entry.isDirectory() && entry.getName().endsWith("pom.properties")) {
+                                propertiesFiles.add(entry.getName());
+                            }
+                        }
+                    }
+                } catch (URISyntaxException | IOException e) {
+                    throw new PippoRuntimeException("Failed to get classes for package '{}'", packageUrl);
+                }
+            }
+        }
+        return propertiesFiles;
     }
 
 }